--- apps/app/drizzle.config.ts
@@ -1,14 +1,0 @@
-/// <reference types="node" />
-import { defineConfig } from 'drizzle-kit';
-import type { Config } from 'drizzle-kit';
-
-export default defineConfig({
-	schema: './src/lib/database/schema.ts',
-	out: './src/lib/database/migrations',
-	dialect: 'postgresql',
-	dbCredentials: {
-		url: process.env.DATABASE_URL!
-	},
-	verbose: true,
-	strict: true
-} satisfies Config);
