+++ apps/app/package.json
@@ -10,10 +10,6 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write .",
-		"db:generate": "drizzle-kit generate",
-		"db:migrate": "drizzle-kit migrate",
-		"db:push": "drizzle-kit push",
-		"db:studio": "drizzle-kit studio",
 		"auth:generate": "npx @better-auth/cli generate"
 	},
 	"devDependencies": {
@@ -61,6 +57,7 @@
 		"@tiptap/pm": "^2.2.1",
 		"@tiptap/starter-kit": "^2.2.1",
 		"@typyst/auth": "workspace:^",
+		"@typyst/db": "workspace:^",
 		"@typyst/ui": "workspace:^",
 		"better-auth": "^1.0.0",
 		"bits-ui": "^0.15.1",
