--- apps/app/src/lib/database/schema.ts
@@ -1,88 +1,0 @@
-import { bigint, boolean, jsonb, pgTable, text, timestamp } from 'drizzle-orm/pg-core';
-
-// Better Auth tables - these will be managed by Better Auth
-export const user = pgTable('user', {
-	id: text('id').primaryKey(),
-	name: text('name').notNull(),
-	email: text('email').notNull().unique(),
-	emailVerified: boolean('email_verified').notNull().default(false),
-	image: text('image'),
-	createdAt: timestamp('created_at', { withTimezone: true }).defaultNow().notNull(),
-	updatedAt: timestamp('updated_at', { withTimezone: true }).defaultNow().notNull()
-});
-
-export const session = pgTable('session', {
-	id: text('id').primaryKey(),
-	expiresAt: timestamp('expires_at', { withTimezone: true }).notNull(),
-	token: text('token').notNull().unique(),
-	createdAt: timestamp('created_at', { withTimezone: true }).defaultNow().notNull(),
-	updatedAt: timestamp('updated_at', { withTimezone: true }).defaultNow().notNull(),
-	ipAddress: text('ip_address'),
-	userAgent: text('user_agent'),
-	userId: text('user_id')
-		.notNull()
-		.references(() => user.id, { onDelete: 'cascade' })
-});
-
-export const account = pgTable('account', {
-	id: text('id').primaryKey(),
-	accountId: text('account_id').notNull(),
-	providerId: text('provider_id').notNull(),
-	userId: text('user_id')
-		.notNull()
-		.references(() => user.id, { onDelete: 'cascade' }),
-	accessToken: text('access_token'),
-	refreshToken: text('refresh_token'),
-	idToken: text('id_token'),
-	accessTokenExpiresAt: timestamp('access_token_expires_at', { withTimezone: true }),
-	refreshTokenExpiresAt: timestamp('refresh_token_expires_at', { withTimezone: true }),
-	scope: text('scope'),
-	password: text('password'),
-	createdAt: timestamp('created_at', { withTimezone: true }).defaultNow().notNull(),
-	updatedAt: timestamp('updated_at', { withTimezone: true }).defaultNow().notNull()
-});
-
-export const verification = pgTable('verification', {
-	id: text('id').primaryKey(),
-	identifier: text('identifier').notNull(),
-	value: text('value').notNull(),
-	expiresAt: timestamp('expires_at', { withTimezone: true }).notNull(),
-	createdAt: timestamp('created_at', { withTimezone: true }).defaultNow().notNull(),
-	updatedAt: timestamp('updated_at', { withTimezone: true }).defaultNow().notNull()
-});
-
-// User-scoped application tables
-export const collection = pgTable('collection', {
-	path: text('path').primaryKey(),
-	name: text('name').notNull(),
-	lastOpened: timestamp('last_opened', { withTimezone: true }).notNull(),
-	userId: text('user_id').references(() => user.id, { onDelete: 'cascade' }) // Made optional for now
-});
-
-export const collectionSettings = pgTable('collection_settings', {
-	collectionPath: text('collection_path')
-		.primaryKey()
-		.references(() => collection.path),
-	editor: jsonb('editor').notNull(),
-	notes: jsonb('notes').notNull(),
-	userId: text('user_id').references(() => user.id, { onDelete: 'cascade' }) // Made optional for now
-});
-
-export const entry = pgTable('entry', {
-	path: text('path').primaryKey(),
-	name: text('name'),
-	parentPath: text('parent_path').notNull(),
-	collectionPath: text('collection_path').references(() => collection.path),
-	content: text('content'),
-	isFolder: boolean('is_folder').default(false),
-	size: bigint('size', { mode: 'number' }),
-	createdAt: timestamp('created_at', { withTimezone: true }).defaultNow().notNull(),
-	updatedAt: timestamp('updated_at', { withTimezone: true }).defaultNow().notNull(),
-	userId: text('user_id').references(() => user.id, { onDelete: 'cascade' }) // Made optional for now
-});
-
-// Export types for TypeScript
-export type User = typeof user.$inferSelect;
-export type Session = typeof session.$inferSelect;
-export type Collection = typeof collection.$inferSelect;
-export type Entry = typeof entry.$inferSelect;
