+++ apps/web/tailwind.config.js
@@ -1,7 +1,6 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const sharedConfig = require('@typyst/tailwind-config/tailwind.config.js');
+import sharedConfig from '@typyst/tailwind-config/tailwind.config.js';
 
-module.exports = {
+export default {
 	presets: [sharedConfig],
 	darkMode: ['class'],
 	content: ['./src/**/*.{html,js,svelte,ts}', '../../packages/ui/**/*.{html,js,svelte,ts}']
