+++ packages/db/drizzle.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from 'drizzle-kit';
+import type { Config } from 'drizzle-kit';
+
+export default defineConfig({
+  schema: './src/schema/*.ts',
+  out: './migrations/supabase',
+  dialect: 'postgresql',
+  dbCredentials: {
+    url: process.env.DATABASE_URL!
+  },
+  verbose: true,
+  strict: true
+} satisfies Config);
